Endpoints updates and some DB fixes.

+++ src/main/java/book/storage/app/model/Book.java
@@ -14,7 +14,7 @@ public class Book  {
     @Id
     @Column(name = "id", unique = true, nullable = false, updatable = false)
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @Column(name = "name", length = 50, nullable = false)
     private String name;
@@ -38,9 +38,9 @@ public class Book  {
 
     public Book() { }
 
-    public Long getId() { return id; }
+    public Integer getId() { return id; }
 
-    public void setId(Long id) { this.id = id; }
+    public void setId(Integer id) { this.id = id; }
 
     public String getName() {
         return name;
